Make nicos E4-compatible
