+++ base/uk.ac.stfc.isis.ibex.ui.nicos/build.properties
@@ -3,4 +3,3 @@ output.. = bin/
 bin.includes = META-INF/,\
                .,\
                icons/,\
-               plugin.xml
