--- base/uk.ac.stfc.isis.ibex.ui.nicos/plugin.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.4"?>
-<plugin>
-   <extension
-         point="org.eclipse.ui.perspectives">
-         <perspective
-            class="uk.ac.stfc.isis.ibex.ui.nicos.Perspective"
-            id="uk.ac.stfc.isis.ibex.ui.nicos.perspective"
-            name="NICOS Scripting">
-      </perspective>
-   </extension>
-
-	<extension
-         point="org.eclipse.ui.perspectiveExtensions">
-            <perspectiveExtension
-            targetID="uk.ac.stfc.isis.ibex.ui.nicos.perspective">
-         <view
-               closeable="false"
-               id="uk.ac.stfc.isis.ibex.ui.nicos.nicosview"
-               minimized="false"
-               moveable="false"
-               ratio="0.1f"
-               relationship="right"
-               relative="uk.ac.stfc.isis.ibex.ui.perspectives.PerspectiveSwitcher"
-               showTitle="false"
-               standalone="true"
-               visible="true">
-         </view>
-      </perspectiveExtension>
-   </extension>
-   <extension
-         point="org.eclipse.ui.views">
-         <view
-            allowMultiple="false"
-            class="uk.ac.stfc.isis.ibex.ui.nicos.NicosView"
-            id="uk.ac.stfc.isis.ibex.ui.nicos.nicosview"
-            name="NICOS Scripting"
-            restorable="true">
-      </view>
-   </extension>
-
-</plugin>
