+++ base/uk.ac.stfc.isis.ibex.ui.weblinks/plugin.xml
@@ -11,10 +11,4 @@
             restorable="true">
       </view>
    </extension>
-   <extension
-         point="uk.ac.stfc.isis.ibex.instrument.info">
-      <InstrumentInfoReceiver
-            class="uk.ac.stfc.isis.ibex.ui.weblinks.WebLinksInstrumentSwitcher">
-      </InstrumentInfoReceiver>
-   </extension>
 </plugin>
